+++ BearishEngine/src/Core/BEMFile.h
@@ -214,7 +214,7 @@ namespace Bearish {
 			WriteBytes(output, (u8*)&numVertices, 1, sizeof(u32));
 			WriteBytes(output, (u8*)&numIndices, 1, sizeof(u32));
 			WriteBytes(output, (u8*)&skinned, 1, sizeof(u8));
-			u8 nameLength = name.length();
+			u8 nameLength = (u8)name.length();
 			WriteBytes(output, &nameLength, 1, sizeof(u8));
 			WriteBytes(output, (u8*)name.c_str(), 1, sizeof(u8) * nameLength);
 			printf("%d\n", bytesWritten);
